backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
pydantic[email]==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
openai==1.12.0
httpx==0.27.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
email-validator==2.1.0.post1
pytest==8.0.2
pytest-asyncio==0.23.5
pandas==2.2.2
cryptography==42.0.5
gunicorn==21.2.0